# Use `execute_batch2` rather than `execute_batch` to fix performance regression for fixture loading