--- a/railties/test/generators/db_system_change_generator_test.rb
+++ b/railties/test/generators/db_system_change_generator_test.rb
@@ -68,7 +68,7 @@ module Rails
 
             assert_file("Gemfile") do |content|
               assert_match "# Use sqlite3 as the database for Active Record", content
-              assert_match "gem 'sqlite3', '~> 1.3', '>= 1.3.6'", content
+              assert_match "gem 'sqlite3', '~> 1.4'", content
             end
           end